refactor: move layout_chunking_config from chunking directory to layout_handler directory
