+++ src/ingestion_pipeline/chunking/strategies/table/layout_table.py
@@ -7,7 +7,7 @@ from textractor.entities.layout import Layout, Line
 from textractor.entities.table import Table
 
 from ingestion_pipeline.chunking.exceptions import ChunkException
-from ingestion_pipeline.chunking.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
 from ingestion_pipeline.chunking.strategies.base import ChunkingStrategyHandler
 from ingestion_pipeline.chunking.strategies.table.base import BaseTableChunker
