+++ tests/chunking/strategies/table/test_line_chunker.py
@@ -8,7 +8,7 @@ import pytest
 # Application modules that we will be patching
 import ingestion_pipeline.chunking.strategies.table.base as base_module
 import ingestion_pipeline.chunking.strategies.table.line_chunker as line_chunker_module
-from ingestion_pipeline.chunking.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.table.line_chunker import LineTableChunker
