+++ tests/chunking/strategies/test_layout_text_strategy.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock, call
 import pytest
 
 import ingestion_pipeline.chunking.strategies.layout_text as layout_text_module
-from ingestion_pipeline.chunking.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.strategies.layout_text import LayoutTextChunkingStrategy
